pyproject.toml
[tool.poetry]
name = "external-metadata-awareness"
version = "0.1.0"
description = "Tools for fetching and processing external schemas, ontologies and metadata records. Includes value set maintenance."
authors = ["Mark A. Miller <MAM@lbl.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
pymongo = "^4.8.0"
tqdm = "^4.66.5"
lxml = "^5.3.0"
oaklib = "^0.6.14"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
list-mongodb-paths = 'external_metadata_awareness.list_mongodb_paths:main'
xml-to-mongo = 'external_metadata_awareness.xml_to_mongo:process_xml_with_progress'

